requirements.txt
pyyaml
snowflake-snowpark-python
pandas
streamlit